requirements.txt
pypdf
python-docx
sentence-transformers
scikit-learn
numpy
transformers
torch
accelerate
google-generativeai
python-dotenv
pdfplumber
PyMuPDF
pytesseract
Pillow
docx2txt
langdetect
beautifulsoup4
pandas
tqdm
unidecode
fuzzywuzzy
python-Levenshtein
chromadb
streamlit
plotly
openpyxl